Check if candwin is executable prior to trying to launch it
Due to QTBUG-5990, a failed QProcess::start will leave behind a
defunct (zombified) process. This can be difficult to notice at
first, and since CandidateWindowProxy::initializeProcess() is
called quite often, a large number of defunct processes can accumulate,
eventually exhausting machine resources and forcing a reboot.

This patch adds code to check if the candwin binary exists and is
executable before attempting to launch it. A warning is printed
if the binary doesn't exist, in an attempt to tell the user something
is wrong. This check is of course not bulletproof, but it is something.